stage13_skills/pyproject.toml
[project]
name = "stage13-skills"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "decorator>=5.2.1",
    "deepagents>=0.5.1",
    "dotenv>=0.9.9",
    "faster-whisper>=1.2.1",
    "funasr>=1.3.1",
    "ipdb>=0.13.13",
    "langchain-modal>=0.0.3",
    "langchain-openai>=1.1.12",
    "modal>=1.4.1",
    "pyannote-audio>=4.0.4",
]
